backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
supabase==2.10.0
redis==5.2.1
paho-mqtt==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
anyio==4.7.0
